requirements.txt
pyrogram
tgcrypto
mysql-connector-python
python-dotenv
pymongo
deep-translator